Add support for randomizations in expectations
