+++ lib/mumuki/domain/extensions.rb
@@ -1,5 +1,7 @@
 require_relative './extensions/string'
+require_relative './extensions/symbol'
 require_relative './extensions/array'
 require_relative './extensions/module'
 require_relative './extensions/hash'
+require_relative './extensions/nil'
 require_relative './extensions/time'
