+++ lib/mumuki/domain/extensions/nil.rb
@@ -1,0 +1,17 @@
+class NilClass
+  def affable
+  end
+
+  def markdownified(**options)
+  end
+
+  def sanitized
+  end
+
+  def randomize_with(randomizer, seed)
+  end
+
+  def insert_last(element)
+    [element]
+  end
+end
