+++ lib/mumuki/domain/extensions/symbol.rb
@@ -1,0 +1,5 @@
+class Symbol
+  def randomize_with(randomizer, seed)
+    self.to_s.randomize_with(randomizer, seed).to_sym
+  end
+end
